+++ src/front/js/layout.js
@@ -11,7 +11,7 @@ import { ProfileAdmon } from "./pages/profileadmon.js";
 import { Demo } from "./pages/demo";
 import { Single } from "./pages/single";
 import { PerfilEspecialista} from "./pages/perfilEspecialista";
-import { Register } from "./pages/register"; // Importa el componente de registro que creaste
+import { Register } from "./pages/registerespecialist.js"; 
 
 import injectContext from "./store/appContext";
 
@@ -40,7 +40,7 @@ const Layout = () => {
                         <Route element={<DatosAdmon />} path="/admons" />
                         <Route element={<Demo />} path="/demo" />
                         <Route element={<Single />} path="/single/:theid" />
-                        <Route element={<Register />} path="/register" /> 
+                        <Route element={<Register />} path="/registerespicialist" /> 
                         <Route element={<PerfilEspecialista />} path="/perfilEspecialista" /> 
                         <Route element={<h1>Not found!</h1>} />
                         <Route element={<CrearProyecto />} path="/nuevoproyecto" />
